requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
httpx>=0.25.2
python-telegram-bot==20.7
SQLAlchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
schedule==1.2.1
pytz==2023.3
jinja2==3.1.2
aiohttp==3.9.1
aiosqlite==0.19.0
alembic==1.13.1 